- Add: DDL storage functionality   In some of the performance counters, get top 10, or all rows (DBName, ObjectName) and get various   information from the ASE Server (sp_help/procText, sp__optdiag, sp_depends) on tables/procedures.   If the table or procedure has dependencies to other tables/procedures, add q DDL request of those objects as well.   Next step: make a GUI so the values can be viewed
- Added functionality to the send info subsystem.

- a bunch of small fixes

Note: this is still a intermediate check in, a bunch of thing need to be done before next release

+++ src/com/asetune/check/server/db_cleanup.php
@@ -23,11 +23,66 @@
 <BODY>
 
 <?php
+	//------------------------------------------
+	// Below is properties sent by the client, vstuff them into local variables
+	//------------------------------------------
+	$doAction = $_GET['doAction'];
+
 	function doCleanup($sql)
 	{
 		echo "EXEC: <code>$sql</code><br>\n";
 		mysql_query($sql) or die("ERROR: " . mysql_error());
 		printf("Records affected: %d<br>\n", mysql_affected_rows());
+		printf("<br>\n");
+	}
+
+	function describe($tabname)
+	{
+		echo "DESCRIBE: <code>$tabname</code><br>\n";
+
+		$sql = "DESCRIBE $tabname";
+		$result = mysql_query($sql) or die("ERROR: " . mysql_error());
+		if (!$result) {
+			echo mysql_errno() . ": " . mysql_error() . "<br>";
+			die("ERROR: Query to show fields from table failed");
+		}
+
+		$fields_num = mysql_num_fields($result);
+
+		// printing table header
+		echo "<table border=\"0\" class=\"sortable\">";
+		echo "<tr>";
+		// printing table headers
+		for($i=0; $i<$fields_num; $i++)
+		{
+			$field = mysql_fetch_field($result);
+			echo "<td nowrap>{$field->name}</td>";
+		}
+		echo "</tr>\n";
+
+		// printing table rows
+		while($row = mysql_fetch_row($result))
+		{
+			echo "<tr>";
+
+			// $row is array... foreach( .. ) puts every element
+			// of $row to $cell variable
+			$col=-1;
+			foreach($row as $cell)
+			{
+				$col++;
+				$colname = mysql_field_name($result, $col);
+
+				//$cellCont = nl2br($cell, false);
+				//echo "<td nowrap>$cellCont</td>";
+				echo "<td nowrap>$cell</td>";
+			}
+
+			echo "</tr>\n";
+		}
+		echo "</table>\n";
+		mysql_free_result($result);
+
 		printf("<br>\n");
 	}
 
@@ -40,26 +95,79 @@
 
 	echo "<h1>Connected to database</h1>\n";
 
-	doCleanup("delete from asemon_usage where user_name = 'rlarsson'");
-	doCleanup("delete from asemon_usage where user_name = 'gorans'");
-	doCleanup("delete from asemon_usage where user_name = ''");
+	if ( $doAction == "cleanup" )
+	{
+		doCleanup("delete from asemon_usage where user_name = 'rlarsson'");
+		doCleanup("delete from asemon_usage where user_name = 'gorans'");
+		doCleanup("delete from asemon_usage where user_name = ''");
 
-	doCleanup("delete from asemon_usage              where user_name = 'gorans' ");
-	doCleanup("delete from asemon_connect_info       where userName  = 'gorans' ");
-	doCleanup("delete from asemon_udc_info           where userName  = 'gorans' ");
-	doCleanup("delete from asemon_counter_usage_info where userName  = 'gorans' ");
-	doCleanup("delete from asemon_error_info         where userName  = 'gorans' ");
+		doCleanup("delete from asemon_usage              where user_name = 'gorans' ");
+		doCleanup("delete from asemon_connect_info       where userName  = 'gorans' ");
+		doCleanup("delete from asemon_udc_info           where userName  = 'gorans' ");
+		doCleanup("delete from asemon_counter_usage_info where userName  = 'gorans' ");
+		doCleanup("delete from asemon_error_info         where userName  = 'gorans' ");
 
-	// doCleanup("delete from asemon_usage              where user_name = 'gorans' or user_name = 'sybase'");
-	// doCleanup("delete from asemon_udc_info");
-	// doCleanup("update asemon_usage set clientAsemonVersion = '2.1.0.1.dev' where clientAsemonVersion = '2.1.1.dev'");
+		// doCleanup("delete from asemon_usage              where user_name = 'gorans' or user_name = 'sybase'");
+		// doCleanup("delete from asemon_udc_info");
+		// doCleanup("update asemon_usage set clientAsemonVersion = '2.1.0.1.dev' where clientAsemonVersion = '2.1.1.dev'");
 
-	// doCleanup("alter table asemon_connect_info add column usePcs     varchar(5)");
-	// doCleanup("alter table asemon_connect_info add column pcsConfig  varchar(400)");
+		echo "<i><b>--- END OF COMMANDS ---</b></i>\n";
+	}
+	else if ( $doAction == "dbmaint" )
+	{
+		// doCleanup("alter table asemon_connect_info add column usePcs     varchar(5)");
+		// doCleanup("alter table asemon_connect_info add column pcsConfig  varchar(400)");
 
-	// doCleanup("ALTER TABLE asemon_error_info DROP PRIMARY KEY");
-	// doCleanup("ALTER TABLE asemon_error_info ADD PRIMARY KEY (checkId, sendCounter, serverAddTime)");
+		// doCleanup("ALTER TABLE asemon_error_info DROP PRIMARY KEY");
+		// doCleanup("ALTER TABLE asemon_error_info ADD PRIMARY KEY (checkId, sendCounter, serverAddTime)");
+		// doCleanup("ALTER TABLE asemon_error_info MODIFY logMessage varchar(4096)");
+		// doCleanup("ALTER TABLE asemon_usage ADD callerIpAddress varchar(20)");
 
+		// doCleanup("ALTER TABLE asemon_connect_info       ADD COLUMN connectId INT NOT NULL DEFAULT -1");
+		// doCleanup("ALTER TABLE asemon_connect_info       DROP PRIMARY KEY");
+		// doCleanup("ALTER TABLE asemon_connect_info       ADD  PRIMARY KEY (checkId, connectId, serverAddTime)");
+
+		// doCleanup("ALTER TABLE asemon_counter_usage_info ADD COLUMN connectId INT NOT NULL DEFAULT -1");
+		// doCleanup("ALTER TABLE asemon_counter_usage_info DROP PRIMARY KEY");
+		// doCleanup("ALTER TABLE asemon_counter_usage_info ADD  PRIMARY KEY (checkId, connectId, clientTime, cmName)");
+
+		// doCleanup("ALTER TABLE asemon_connect_info       MODIFY connectId INT NOT NULL  DEFAULT -1  AFTER isClusterEnabled");
+		// doCleanup("ALTER TABLE asemon_counter_usage_info MODIFY connectId INT NOT NULL  DEFAULT -1  AFTER userName");
+
+		// doCleanup("ALTER TABLE asemon_connect_info       MODIFY srvIpPort varchar(60)");
+
+
+		// doCleanup("ALTER TABLE asemon_counter_usage_info CHANGE clientTime sessionStartTime TIMESTAMP");
+		// doCleanup("ALTER TABLE asemon_counter_usage_info ADD sessionType      varchar(10)     AFTER serverAddTime");
+		// doCleanup("ALTER TABLE asemon_counter_usage_info ADD sessionEndTime   timestamp NULL  AFTER sessionStartTime");
+
+		// doCleanup("ALTER TABLE asemon_usage MODIFY callerIpAddress varchar(20) AFTER clientCanonicalHostName");
+
+		// doCleanup("ALTER TABLE asemon_usage ADD user_home varchar(50) AFTER user_name");
+
+
+		echo "<i><b>--- END OF COMMANDS ---</b></i>\n";
+	}
+	else if ( $doAction == "check" )
+	{
+		describe("asemon_usage");
+		describe("asemon_connect_info");
+		describe("asemon_udc_info");
+		describe("asemon_counter_usage_info");
+		describe("asemon_error_info");
+
+		echo "<i><b>--- END OF COMMANDS ---</b></i>\n";
+	}
+	else
+	{
+		echo "<h1>UNSUPPORTED COMMAND '" . $doAction . "'</h1>\n";
+		echo "Commands:<br>\n";
+		echo "<ul>\n";
+		echo "  <li><code>cleanup</code> - Delete gorans from various tables</li>\n";
+		echo "  <li><code>dbmaint</code> - Do some DB maintenace command, needs to be coded in the PHP script. Usually this section will be empty</li>\n";
+		echo "  <li><code>check  </code> - Do 'DESCRIBE' on all known tables.</li>\n";
+		echo "</ul>\n";
+	}
 
 	//------------------------------------------
 	// Close connection to the database
